# openapi.yaml
openapi: 3.0.3

info:
  title: User Management API
  description: |
    User account creation, authentication, and password management microservice.

    Password-reset tokens are returned in the response body (in production they would be delivered by email).
  version: 0.1.0

servers:
  - url: http://localhost:8081
    description: Local development

paths:
  /health:
    get:
      summary: Health check
      operationId: health
      tags: [System]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Service and database are healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: ok
                database: ok
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          description: Database is unreachable or degraded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: degraded
                database: unavailable

  /users:
    post:
      summary: Create a new user account
      operationId: createUser
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            example:
              email: alice@example.com
              username: alice
              password: s3cretP@ss
      responses:
        '200':
          description: User created (inactive). Confirm email to activate the account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
              example:
                message: Account created. Use the confirmation token to activate your account.
                confirmation_token: a3f1c2d4e5b6...
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'

  /users/{id}/password:
    put:
      summary: Change a user's password
      operationId: changePassword
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the user whose password is being changed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /auth/login:
    post:
      summary: Authenticate and receive a JWT
      operationId: login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: alice@example.com
              password: s3cretP@ss
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /auth/confirm-email:
    get:
      summary: Confirm email address via link click
      description: |
        Activates the user account associated with the confirmation token. Intended
        for use by email clients that fire a GET request when the user clicks the
        confirmation link. Idempotent: calling it a second time returns 401 (token
        already cleared after first activation).
      operationId: confirmEmailGet
      tags: [Auth]
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: 64-character hex confirmation token from the registration email
      responses:
        '204':
          description: Account activated successfully
        '401':
          description: Token not found, already used, or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Confirm email address and activate account
      description: |
        Activates the user account associated with the confirmation token returned
        by `POST /users`. Idempotent: calling it a second time after activation
        returns 401 (token already cleared).
      operationId: confirmEmail
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmEmailRequest'
            example:
              token: a3f1c2d4e5b6...
      responses:
        '204':
          description: Account activated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Token not found, already used, or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  /auth/password-reset/request:
    post:
      summary: Request a password-reset token
      description: |
        Always returns `200 OK` regardless of whether the email exists, to prevent
        user enumeration. If the email is registered the reset token is included in
        the response body; otherwise a generic message is returned.
      operationId: requestPasswordReset
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
            example:
              email: alice@example.com
      responses:
        '200':
          description: Request processed (email may or may not be registered)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PasswordResetTokenResponse'
                  - $ref: '#/components/schemas/GenericMessageResponse'
              examples:
                email_found:
                  summary: Email was registered
                  value:
                    reset_token: a3f1c2d4e5b6...
                email_not_found:
                  summary: Email was not registered
                  value:
                    message: If the email is registered you will receive a reset token
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /auth/password-reset/confirm:
    post:
      summary: Complete a password reset
      operationId: confirmPasswordReset
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
            example:
              token: a3f1c2d4e5b6...
              new_password: fresh_P@ssw0rd
      responses:
        '204':
          description: Password reset successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Token is expired, already used, or not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ok, degraded]
        database:
          type: string
          enum: [ok, error, unavailable]

    CreateUserRequest:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 8

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: Signed HS256 JWT valid for `JWT_TTL_HOURS` hours
        user:
          $ref: '#/components/schemas/UserResponse'
        roles:
          type: array
          items:
            type: string
          description: Role names assigned to the user
        permissions:
          type: array
          items:
            type: string
          description: Permission names granted via the user's roles

    ChangePasswordRequest:
      type: object
      required: [new_password]
      properties:
        current_password:
          type: string
          description: Required when the caller is changing their own password; omit when an admin changes another user's password
        new_password:
          type: string
          minLength: 8

    CreateUserResponse:
      type: object
      properties:
        message:
          type: string
        confirmation_token:
          type: string
          description: 64-character hex token; use with /auth/confirm-email

    ConfirmEmailRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: 64-character hex token from the registration response

    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email

    PasswordResetConfirm:
      type: object
      required: [token, new_password]
      properties:
        token:
          type: string
          description: 64-character hex token from the reset request
        new_password:
          type: string
          minLength: 8

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    PasswordResetTokenResponse:
      type: object
      properties:
        reset_token:
          type: string
          description: 64-character hex token; use with /auth/password-reset/confirm

    GenericMessageResponse:
      type: object
      properties:
        message:
          type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string

  responses:
    BadRequest:
      description: Invalid request body or validation failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: password must be at least 8 characters

    Unauthorized:
      description: Invalid credentials or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: invalid credentials

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: user not found

    Conflict:
      description: Email or username already taken
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: email or username already taken

    Forbidden:
      description: Caller lacks the required permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: insufficient permissions

    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'database error: ...'